Support for 'prompt'.
